Implementation of child Labour Version 1
This commit has the following changes:

The dialouge for the Child Labour has been completed. It has been added to the red cube named Makelele under componments ( Named NPC conversion) The prefab, Makelele, has been created and has been put in the Prefab folder.

Things to note:

Scene is created with a ground base and a red cube only.

I cannot open the dialouge editor for some reason. If someone gets the chance please try to open the dialouge editor and see if you get the dialouge editor window.