Added Journal Scripts and Testable Scene
The Journal works just like the inventory
Click the "Journal" button to open inventory. click again to close.
Press "K" on the keyboard to get key information. The information will be displayed in the journal.
pressing k again will display that the info was already found. This uses the interaction condition collection and a bool that is set to true along side getting the key info.
Possible next step is to have the journal script check if info was already noted and not repeat the note, removing the steps to make a condition and reaction etc etc.