First changes to conversation
Add NavMovementReaction to move NPCs
Changed the test dialogue to have events to change cameras and move NPCs mid conversation.
Added ToonShader matrial to placeholder game objects that represent NPCs.
Add DummyConversation to Conversation Manager prefab and edited Converstation Manager script to play and end that converstation. This is to get past the bug that makes the first conversation UI appear in a scene to be simi-transparent.